requirements.txt
#python -m venv venv
#.\venv\Scripts\activate WINDOWS
# source ./venv/bin/activate LINUX
#pip install -r requirements.txt
tinkerforge == 2.1.30
Pillow 
customtkinter 
pyinstaller

# for Executable
# pyinstaller --noconfirm --onedir --windowed --add-data "C:/Users/cybwa/AppData/Local/Programs/Python/Python311/Lib/site-packages/customtkinter;customtkinter/"  "Steuerung.py"
